services/api/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sidetrack-api"
version = "0.1.0"
dependencies = [
    "fastapi==0.111.1",
    "uvicorn[standard]==0.30.1",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "python-dotenv==1.0.1",
    "psycopg[binary]==3.2.1",
    "SQLAlchemy==2.0.32",
    "alembic==1.13.2",
    "requests==2.32.3",
    "httpx[async]==0.27.0",
    "tenacity==8.2.3",
    "passlib[bcrypt]==1.7.4",
    "redis==5.0.7",
    "rq==1.16.2",
    "structlog==24.1.0",
    "opentelemetry-api==1.36.0",
    "opentelemetry-sdk==1.36.0",
    "opentelemetry-instrumentation-fastapi==0.57b0",
    "opentelemetry-instrumentation-httpx==0.57b0",
    "opentelemetry-instrumentation-sqlalchemy==0.57b0",
    "pytest-asyncio==0.23.7",
]

[tool.setuptools]
packages = ["app"]
include-package-data = true

[tool.setuptools.package-data]
"app" = ["model_data/*.json"]
